Bugfixes (per sonar report)

+++ jpo-ode-plugins/src/main/java/us/dot/its/jpo/ode/plugin/j2735/oss/OssHeight.java
@@ -12,7 +12,7 @@ public class OssHeight {
 	        throw new IllegalArgumentException("Vehicle height out of bounds");
 	    }
 	    
-		return BigDecimal.valueOf(height.intValue() * 5, 2);
+		return BigDecimal.valueOf(height.intValue() * (long)5, 2);
 	}
 
 }
